[callbacks] optimized callbacks
 - Removed the union of C/JS structs so callbacks only do 1 allocation
 - Changed callback ID's to int16_t's instead of int32_t's
 - Changed all modules using calbacks to use int16_t's

+++ src/zjs_aio.c
@@ -19,7 +19,7 @@ static struct nano_sem aio_sem;
 
 typedef struct aio_handle {
     jerry_value_t pin_obj;
-    int32_t callback_id;
+    int16_t callback_id;
     double value;
     jerry_value_t jvalue;
 } aio_handle_t;
